+++ zinc/parser/zincLexer.py
@@ -1,4 +1,4 @@
-# Generated from zinc/parser/zinc.g4 by ANTLR 4.13.2
+# Generated from zinc.g4 by ANTLR 4.13.2
 from antlr4 import *
 from io import StringIO
 import sys